requirements.txt
# Core Data Processing
pandas==2.1.3
numpy==1.24.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Validation and Quality
great-expectations==0.18.3
cerberus==1.3.5

# API and Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and Monitoring
structlog==23.2.0

# Data Visualization (for reports)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# AWS SDK (optional for cloud integration)
boto3==1.34.0
awscli==1.32.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development Tools
black==23.11.0
flake8==6.1.0
isort==5.12.0 